This seems to fix the checkbox rendering.

+++ source/Xamarin.Forms.Platform.AvaloniaUI/Implementation/Controls/RoundButton.cs
@@ -19,7 +19,7 @@ public class RoundButton : Avalonia.Controls.Button, IStyleable
     {
         get
         {
-            return (int)GetValue(CornerRadiusProperty);
+            return GetValue(CornerRadiusProperty);
         }
         set
         {
